Clean up routes
- Delete the following from backend/routes:
  - comments.js
  - genres.js
- Add backend/routes/middleware/singleShowMiddleware.js
- Changes to singleShow.js
  - remove GET all comments
  - change GET single show for single user route
    - added new middleware functions from the above
- Add TODOs to backend/server.js
- Other small changes
