+++ backend/routes/singleShow.js
@@ -8,59 +8,28 @@
  * `/show/:id` : POST new comment
  *
  * Status:
- * `/show/:id` : GET requests NOT complete
+ * `/show/:id` : GET requests complete
  * `/show/:id` : POST request comlete
  */
 const express = require('express');
-const Show = require('../models/show');
 const Comment = require('../models/comment');
+
+const middleware = require('./middleware/singleShowMiddleware');
 
 const { body, validationResult } = require('express-validator/check');
 
 const router = express.Router();
 
-// TODO: split this into two middleware functions and unify it later
 /**
- * GET single show for a single user
+ * Middleware functions
  */
-router.get('/:id', async (req, res) => {
-  try {
-    const queryShows = await Show.findOne({
-      where: {
-        id: req.params.id
-      }
-    });
-    if(queryShows == null || queryShows.length === 0) {
-      // TODO flesh this out more
-      res.status(404).send();
-    } else {
-      const singleShow = queryShows.dataValues;
-      res.json({show: singleShow});
-    }
-  } catch(error) {
-    console.log(error);
-    res.status(500).send();
-  }
-});
+const { getCommentsForShow, getSingleShow } = middleware;
+
 
 /**
- * GET all comments
+ * GET single show for a single user and comments
  */
-router.get('/:id', async (req, res) => {
-  try {
-    const queryComments = await Comment.findAll();
-    const comments = queryComments.map(comment => {
-      return {
-        id: comment.dataValues.id,
-        user_id: comment.dataValues.user_id,
-        show_id: comment.dataValues.show_id,
-        comment_body: comment.dataValues.comment_body
-      }});
-    res.json({comments: comments});
-  } catch(error) {
-    res.status(500).send();
-  }
-});
+router.get('/:id', getSingleShow, getCommentsForShow);
 
 /**
  * POST new comment
